AStarPathfinding - add some changes
This code implements the A * algorithm to find the shortest path in a two-dimensional maze. It uses heuristics to estimate the distance to the endpoint and chooses the path that seems most promising. In this example, the maze is represented as a grid where some cells are locked (cannot be traversed). The algorithm starts at the starting point and looks for a path to the end point, bypassing obstacles.